requirements.txt
anyio==4.3.0
certifi==2024.2.2
coverage==7.4.4
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
pendulum==3.0.0
pluggy==1.4.0
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
time-machine==2.14.1
tinydb==4.8.0
tzdata==2024.1
